Moved input filter chain to separate module
